+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+0.0.7
+----
+* GPIO edge detection expanded to include AP-EINT1 and AP-EINT3 as those are the only other pins that support edge detection
+
+0.0.6
+----
+* Initial PWM
+* GPIO edge detection and callback for XIO-P0 to XIO-P7 working
+
 0.0.4
 ____
 * Initial Commit